api-python: move the general-purpose utilities into velstor.libutil
